LPS-58793 Apply changes to fit the environment
