+++ modules/apps/wiki/wiki-service/src/com/liferay/wiki/exportimport/portlet/preferences/processor/WikiDisplayExportImportPortletPreferencesProcessor.java
@@ -17,6 +17,12 @@ package com.liferay.wiki.exportimport.portlet.preferences.processor;
 import com.liferay.exportimport.portlet.preferences.processor.Capability;
 import com.liferay.exportimport.portlet.preferences.processor.ExportImportPortletPreferencesProcessor;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.dao.orm.DynamicQuery;
+import com.liferay.portal.kernel.dao.orm.Property;
+import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.StringPool;
@@ -26,6 +32,7 @@ import com.liferay.portlet.exportimport.lar.StagedModelDataHandlerUtil;
 import com.liferay.wiki.constants.WikiPortletKeys;
 import com.liferay.wiki.model.WikiNode;
 import com.liferay.wiki.model.WikiPage;
+import com.liferay.wiki.service.WikiPageLocalServiceUtil;
 import com.liferay.wiki.service.permission.WikiResourcePermissionChecker;
 import com.liferay.wiki.service.persistence.WikiNodeUtil;
 
@@ -33,6 +40,7 @@ import java.util.List;
 import java.util.Map;
 
 import javax.portlet.PortletPreferences;
+import javax.portlet.ReadOnlyException;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -62,6 +70,8 @@ public class WikiDisplayExportImportPortletPreferencesProcessor
 			PortletDataContext portletDataContext,
 			PortletPreferences portletPreferences)
 		throws PortletDataException {
+
+		String portletId = portletDataContext.getPortletId();
 
 		long nodeId = GetterUtil.getLong(
 			portletPreferences.getValue("nodeId", StringPool.BLANK));
@@ -96,8 +106,14 @@ public class WikiDisplayExportImportPortletPreferencesProcessor
 			return portletPreferences;
 		}
 
-		portletDataContext.addPortletPermissions(
-			WikiResourcePermissionChecker.RESOURCE_NAME);
+		try {
+			portletDataContext.addPortletPermissions(
+				WikiResourcePermissionChecker.RESOURCE_NAME);
+		}
+		catch (PortalException pe) {
+			throw new PortletDataException(
+				"Unable to export portlet permissions", pe);
+		}
 
 		StagedModelDataHandlerUtil.exportReferenceStagedModel(
 			portletDataContext, portletId, node);
@@ -106,7 +122,13 @@ public class WikiDisplayExportImportPortletPreferencesProcessor
 			getPageActionableDynamicQuery(
 				portletDataContext, node.getNodeId(), portletId);
 
-		actionableDynamicQuery.performActions();
+		try {
+			actionableDynamicQuery.performActions();
+		}
+		catch (PortalException pe) {
+			throw new PortletDataException(
+				"Unable to export referenced pages", pe);
+		}
 
 		return portletPreferences;
 	}
@@ -117,8 +139,14 @@ public class WikiDisplayExportImportPortletPreferencesProcessor
 			PortletPreferences portletPreferences)
 		throws PortletDataException {
 
-		portletDataContext.importPortletPermissions(
-			WikiResourcePermissionChecker.RESOURCE_NAME);
+		try {
+			portletDataContext.importPortletPermissions(
+				WikiResourcePermissionChecker.RESOURCE_NAME);
+		}
+		catch (PortalException pe) {
+			throw new PortletDataException(
+				"Unable to import portlet permissions", pe);
+		}
 
 		StagedModelDataHandlerUtil.importReferenceStagedModels(
 			portletDataContext, WikiNode.class);
@@ -136,10 +164,62 @@ public class WikiDisplayExportImportPortletPreferencesProcessor
 
 			nodeId = MapUtil.getLong(nodeIds, nodeId, nodeId);
 
-			portletPreferences.setValue("nodeId", String.valueOf(nodeId));
+			try {
+				portletPreferences.setValue("nodeId", String.valueOf(nodeId));
+			}
+			catch (ReadOnlyException roe) {
+				throw new PortletDataException(
+					"Unable to update portlet preferences during import", roe);
+			}
 		}
 
 		return portletPreferences;
 	}
+
+	protected ActionableDynamicQuery getPageActionableDynamicQuery(
+		final PortletDataContext portletDataContext, final long nodeId,
+		final String portletId) {
+
+		ActionableDynamicQuery actionableDynamicQuery =
+			WikiPageLocalServiceUtil.getExportActionableDynamicQuery(
+				portletDataContext);
+
+		final ActionableDynamicQuery.AddCriteriaMethod addCriteriaMethod =
+			actionableDynamicQuery.getAddCriteriaMethod();
+
+		actionableDynamicQuery.setAddCriteriaMethod(
+			new ActionableDynamicQuery.AddCriteriaMethod() {
+
+				@Override
+				public void addCriteria(DynamicQuery dynamicQuery) {
+					addCriteriaMethod.addCriteria(dynamicQuery);
+
+					Property property = PropertyFactoryUtil.forName("nodeId");
+
+					dynamicQuery.add(property.eq(nodeId));
+				}
+
+		});
+
+		actionableDynamicQuery.setPerformActionMethod(
+			new ActionableDynamicQuery.PerformActionMethod() {
+
+				@Override
+				public void performAction(Object object)
+					throws PortalException {
+
+					WikiPage page = (WikiPage)object;
+
+					StagedModelDataHandlerUtil.exportReferenceStagedModel(
+						portletDataContext, portletId, page);
+				}
+
+		});
+
+		return actionableDynamicQuery;
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		WikiDisplayExportImportPortletPreferencesProcessor.class);
 
 }
